fix: remove json transformation
